@@ -23,7 +23,7 @@ namespace MyTeam.Models.Domain
         public DateTime Published { get; set; }
 
         public virtual Member Author { get; set; }
-        public virtual Game Game { get; set; }
+        public virtual Event Game { get; set; }
         public virtual Club Club { get; set; }
         public virtual ICollection<Comment> Comments { get; set; }
 
