@@ -30,9 +30,7 @@ namespace MyTeam.Models.Domain
         public virtual ICollection<MemberTeam> MemberTeams { get; set; }
         public virtual ICollection<Season> Seasons { get; set; }
 
-        public virtual ICollection<Game> Games { get; set; }
+        public virtual ICollection<Event> Games { get; set; }
         public virtual ICollection<EventTeam> EventTeams { get; set; }
-        [NotMapped]
-        public virtual IEnumerable<Event> Trainings => EventTeams?.Select(e => e.Event).Where(e => e.IsTraining);
     }
 }
