@@ -40,7 +40,7 @@ type PeriodStart = DateTime
 
 type GetAttendance = Database -> ClubId -> Year -> SelectedYear * Years * PlayerAttendanceSummary list
 type GetPreviousTrainings = Database -> ClubId -> Event list
-type GetTraining = Database -> EventId -> Event
+type GetTraining = Database -> EventId -> Event option
 type GetPlayers = Database -> ClubId -> EventId -> Playerlist
 type ConfirmAttendance = Database -> ClubId -> EventId ->  MemberId -> CheckboxPayload -> HttpResult<unit>
 type ConfirmVictory = Database -> ClubId -> EventId ->  MemberId -> CheckboxPayload -> HttpResult<unit>
